+++ apps/api/package.json
@@ -9,12 +9,12 @@
     "start": "bun src/index.ts"
   },
   "devDependencies": {
-    "@types/node": "^20.0.0",
-    "esbuild": "^0.21.4",
-    "npm-run-all2": "^6.2.0"
+    "@types/node": "*",
+    "esbuild": "*",
+    "npm-run-all2": "*"
   },
   "dependencies": {
-    "hono": "^4.8.4",
+    "hono": "*",
     "@hebo/db": "workspace:*"
   }
 }
