+++ apps/app/package.json
@@ -7,43 +7,45 @@
     "dev": "cross-env NEXT_PUBLIC_API_URL=http://localhost:3001 next dev --turbopack",
     "dev:local": "next dev --turbopack",
     "lint": "next lint",
-    "start": "next start",
+    "start": "serve out -p 3000",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
     "clean": "rm -rf .next .turbo node_modules"
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.1.14",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@stackframe/stack": "^2.8.12",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.511.0",
-    "next": "15.3.2",
-    "next-themes": "^0.4.6",
-    "posthog-js": "^1.249.2",
-    "posthog-node": "^4.18.0",
-    "react": "^18.3.0",
-    "react-dom": "^18.3.0",
-    "sonner": "^2.0.3",
-    "tailwind-merge": "^3.3.0",
-    "vaul": "^1.1.2"
+    "@radix-ui/react-dialog": "*",
+    "@radix-ui/react-slot": "*",
+    "@stackframe/stack": "*",
+    "buffer": "*",
+    "class-variance-authority": "*",
+    "clsx": "*",
+    "lucide-react": "*",
+    "next": "*",
+    "next-themes": "*",
+    "posthog-js": "*",
+    "posthog-node": "*",
+    "react": "*",
+    "react-dom": "*",
+    "sonner": "*",
+    "tailwind-merge": "*",
+    "vaul": "*",
+    "@hebo/ui": "workspace:*"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^20",
-    "@types/react": "^18.3.0",
-    "@types/react-dom": "^18.3.0",
-    "eslint-config-next": "15.3.2",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^30.0.0-beta.3",
-    "tailwindcss": "^4",
-    "tw-animate-css": "^1.3.2"
+    "@eslint/eslintrc": "*",
+    "@tailwindcss/postcss": "*",
+    "@testing-library/jest-dom": "*",
+    "@testing-library/react": "*",
+    "@testing-library/user-event": "*",
+    "@types/jest": "*",
+    "@types/node": "*",
+    "@types/react": "*",
+    "@types/react-dom": "*",
+    "eslint-config-next": "*",
+    "jest": "*",
+    "jest-environment-jsdom": "*",
+    "serve": "*",
+    "tailwindcss": "*"
   }
 }
