+++ apps/app/src/app/page.tsx
@@ -1,11 +1,7 @@
-import { stackServerApp } from "../stack";
 import { Footer } from "@/components/auth/Footer";
-import HomeContent from "@/components/HomeContent";
+import HomeContent from "@/components/common/HomeContent";
 
 export default async function Home() {
-  // Protect the page – redirect to /signup if not authenticated
-  await stackServerApp.getUser({ or: "redirect" });
-
   return (
     <div className="flex min-h-screen flex-col">
       <main className="flex flex-1 items-center justify-center p-8 pb-20 sm:p-20">
