+++ apps/app/src/components/auth/AuthLayout.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { ReactNode } from "react";
 import Image from "next/image";
 import Link from "next/link";
