+++ apps/app/src/components/auth/DocsButton.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { BookCheck } from "lucide-react";
 import { Button } from "@/components/ui/button";
