+++ apps/app/src/components/auth/Footer.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import UserButtonWrapper from "./UserButtonWrapper";
