+++ apps/app/src/components/common/Logo.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import Image from "next/image";
 
 interface LogoProps {
