+++ apps/app/src/components/common/NewButton.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { Button } from "@/components/ui/button";
 
 export function NewButton() {
