+++ apps/app/src/components/ui/button.tsx
@@ -1,61 +1,49 @@
 import * as React from "react"
-import { Slot } from "@radix-ui/react-slot"
 import { cva, type VariantProps } from "class-variance-authority"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@hebo/ui"
+import {
+  Button as ShadcnButton,
+  buttonVariants as shadcnButtonVariants,
+} from "@hebo/ui/shadcn/button"
 
-const buttonVariants = cva(
-  "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
-  {
-    variants: {
-      variant: {
-        default:
-          "bg-primary text-primary-foreground shadow-xs hover:bg-primary/90",
-        destructive:
-          "bg-destructive text-white shadow-xs hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60",
-        outline:
-          "border bg-background shadow-xs hover:bg-accent hover:text-accent-foreground dark:bg-input/30 dark:border-input dark:hover:bg-input/50",
-        secondary:
-          "bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80",
-        tertiary:
-          "bg-tertiary text-tertiary-foreground shadow-xs hover:bg-tertiary/80 rounded-sm",
-        ghost:
-          "hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50",
-        link: "text-primary underline-offset-4 hover:underline",
-      },
-      size: {
-        default: "h-9 px-4 py-2 has-[>svg]:px-3",
-        sm: "h-8 rounded-md gap-1.5 px-3 has-[>svg]:px-2.5",
-        lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
-        icon: "size-9",
-      },
+const extendedButtonVariants = cva("", {
+  variants: {
+    variant: {
+      tertiary:
+        "bg-tertiary text-tertiary-foreground shadow-xs hover:bg-tertiary/80 rounded-sm",
     },
-    defaultVariants: {
-      variant: "default",
-      size: "default",
-    },
-  }
-)
+  },
+})
 
-function Button({
-  className,
-  variant,
-  size,
-  asChild = false,
-  ...props
-}: React.ComponentProps<"button"> &
-  VariantProps<typeof buttonVariants> & {
-    asChild?: boolean
-  }) {
-  const Comp = asChild ? Slot : "button"
-
-  return (
-    <Comp
-      data-slot="button"
-      className={cn(buttonVariants({ variant, size, className }))}
-      {...props}
-    />
-  )
+type ExtendedButtonVariantProps = Omit<
+  VariantProps<typeof shadcnButtonVariants>,
+  "variant"
+> & {
+  variant?: VariantProps<typeof shadcnButtonVariants>["variant"] | "tertiary"
 }
 
-export { Button, buttonVariants }
+type ButtonProps = React.ComponentProps<"button"> & ExtendedButtonVariantProps
+
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, ...props }, ref) => {
+    return (
+      <ShadcnButton
+        ref={ref}
+        size={size}
+        variant={variant === "tertiary" ? "default" : variant}
+        className={cn(
+          variant === "tertiary" &&
+          extendedButtonVariants({
+            variant,
+          }),
+          className,
+        )}
+        {...props}
+      />
+    )
+  },
+)
+Button.displayName = "Button"
+
+export { Button, shadcnButtonVariants as buttonVariants }
