+++ apps/app/tailwind.config.ts
@@ -1,16 +1,12 @@
-import { type Config } from "tailwindcss"
+import type { Config } from "tailwindcss";
+import uiPreset from "@hebo/ui";
 
-export default {
-  theme: {
-    extend: {
-      fontSize: {
-        headline: "var(--text-headline)",
-        xl:       "var(--text-xl)",
-        "xl-sm":  "var(--text-xl-sm)",
-        lg:       "var(--text-lg)",
-        base:     "var(--text-base)",
-        sm:       "var(--text-sm)",
-      },
-    },
-  },
-} satisfies Config
+const config: Config = {
+  content: [
+    "./src/**/*.{js,jsx,ts,tsx}",
+    "../../packages/ui/src/**/*.{js,jsx,ts,tsx}",
+  ],
+  presets: [uiPreset],
+};
+
+export default config;
