+++ infra/stacks/app.ts
@@ -13,7 +13,7 @@ const posthogHost = new sst.Secret("PosthogHost", "fakeValue");
 const heboApp = new sst.aws.Nextjs("HeboApp", {
   path: "apps/app",
   domain: $app.stage === "production" ? "cloud.hebo.ai" : `${$app.stage}.cloud.hebo.ai`,
-  environment: { 
+  environment: {
     NEXT_PUBLIC_API_URL: heboApiUrl,
     NEXT_PUBLIC_STACK_PROJECT_ID: stackProjectId.value,
     NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY: stackPublishableClientKey.value,
