+++ packages/ui/src/globals.css
@@ -1,0 +1,154 @@
+@import "tailwindcss";
+@import "tw-animate-css";
+
+@custom-variant dark (&:is(.dark *));
+
+@theme {
+  /* === Colours === */
+  --color-background-default : linear-gradient(0deg,#F8F9D2 0%,#F5F7FA 100%);
+  --color-foreground-default : #000000;
+  --color-background-secondary: rgba(175, 207, 255, 0.25);
+  --color-foreground-secondary: #333333;
+  --color-background-tertiary: rgba(88, 101, 242, 0.6);
+  --color-foreground-tertiary: #FFFFFF;
+  --color-marketing-500      : #E4EDFB;
+
+  /* === Typography === */
+  --text-headline : 8rem;      /* 128 px desktop */
+  --text-xl       : 2rem;      /* 32 px desktop  */
+  --text-xl-sm    : 1.25rem;   /* 20 px mobile   */
+  --text-lg       : 1.5rem;    /* 24 px          */
+  --text-base     : 0.875rem;  /* 14 px          */
+  --text-sm       : 0.75rem;   /* 12 px          */
+}
+
+/* Runtime variables reused outside Tailwind utilities */
+:root {
+  --font-sans  : var(--font-geist-sans);
+  --font-mono  : var(--font-geist-mono);
+  --radius : 1.125rem; /* 18px in rem */
+  --background : var(--color-background-default);
+  --foreground : var(--color-foreground-default);
+  --card : oklch(1 0 0);
+  --card-foreground : oklch(0.145 0 0);
+  --popover : oklch(1 0 0);
+  --popover-foreground : oklch(0.145 0 0);
+  --primary : oklch(0.205 0 0);
+  --primary-foreground : oklch(0.985 0 0);
+  --secondary : var(--color-background-secondary);
+  --secondary-foreground : var(--color-foreground-secondary);
+  --tertiary : var(--color-background-tertiary);
+  --tertiary-foreground : var(--color-foreground-tertiary);
+  --muted : oklch(0.97 0 0);
+  --muted-foreground : oklch(0.556 0 0);
+  --accent : oklch(0.97 0 0);
+  --accent-foreground : oklch(0.205 0 0);
+  --destructive : oklch(0.577 0.245 27.325);
+  --border : oklch(0.922 0 0);
+  --input : oklch(1 0 0 / 75%);
+  --ring : oklch(0.708 0 0);
+  --chart-1 : oklch(0.646 0.222 41.116);
+  --chart-2 : oklch(0.6 0.118 184.704);
+  --chart-3 : oklch(0.398 0.07 227.392);
+  --chart-4 : oklch(0.828 0.189 84.429);
+  --chart-5 : oklch(0.769 0.188 70.08);
+  --sidebar : oklch(0.985 0 0);
+  --sidebar-foreground : oklch(0.145 0 0);
+  --sidebar-primary : oklch(0.205 0 0);
+  --sidebar-primary-foreground : oklch(0.985 0 0);
+  --sidebar-accent : oklch(0.97 0 0);
+  --sidebar-accent-foreground : oklch(0.205 0 0);
+  --sidebar-border : oklch(0.922 0 0);
+  --sidebar-ring : oklch(0.708 0 0);
+}
+
+body {
+  font-family: var(--font-sans);
+  background: var(--color-background-default);
+  color: var(--color-foreground-default);
+}
+
+@theme inline {
+  --radius-sm : calc(var(--radius) - 4px);
+  --radius-md : calc(var(--radius) - 2px);
+  --radius-lg : var(--radius);
+  --radius-xl : calc(var(--radius) + 4px);
+  --color-background : var(--background);
+  --color-foreground : var(--foreground);
+  --color-card : var(--card);
+  --color-card-foreground : var(--card-foreground);
+  --color-popover : var(--popover);
+  --color-popover-foreground : var(--popover-foreground);
+  --color-primary : var(--primary);
+  --color-primary-foreground : var(--primary-foreground);
+  --color-secondary : var(--secondary);
+  --color-secondary-foreground : var(--secondary-foreground);
+  --color-tertiary : var(--tertiary);
+  --color-tertiary-foreground : var(--tertiary-foreground);
+  --color-muted : var(--muted);
+  --color-muted-foreground : var(--muted-foreground);
+  --color-accent : var(--accent);
+  --color-accent-foreground : var(--accent-foreground);
+  --color-destructive : var(--destructive);
+  --color-border : var(--border);
+  --color-input : var(--input);
+  --color-ring : var(--ring);
+  --color-chart-1 : var(--chart-1);
+  --color-chart-2 : var(--chart-2);
+  --color-chart-3 : var(--chart-3);
+  --color-chart-4 : var(--chart-4);
+  --color-chart-5 : var(--chart-5);
+  --color-sidebar : var(--sidebar);
+  --color-sidebar-foreground : var(--sidebar-foreground);
+  --color-sidebar-primary : var(--sidebar-primary);
+  --color-sidebar-primary-foreground : var(--sidebar-primary-foreground);
+  --color-sidebar-accent : var(--sidebar-accent);
+  --color-sidebar-accent-foreground : var(--sidebar-accent-foreground);
+  --color-sidebar-border : var(--sidebar-border);
+  --color-sidebar-ring : var(--sidebar-ring);
+}
+
+.dark {
+  --background : oklch(0.145 0 0);
+  --foreground : oklch(0.985 0 0);
+  --card : oklch(0.205 0 0);
+  --card-foreground : oklch(0.985 0 0);
+  --popover : oklch(0.205 0 0);
+  --popover-foreground : oklch(0.985 0 0);
+  --primary : oklch(0.922 0 0);
+  --primary-foreground : oklch(0.205 0 0);
+  --secondary : oklch(0.269 0 0);
+  --secondary-foreground : oklch(0.985 0 0);
+  --tertiary : oklch(0.269 0.1 240);
+  --tertiary-foreground : oklch(0.985 0 0);
+  --muted : oklch(0.269 0 0);
+  --muted-foreground : oklch(0.708 0 0);
+  --accent : oklch(0.269 0 0);
+  --accent-foreground : oklch(0.985 0 0);
+  --destructive : oklch(0.704 0.191 22.216);
+  --border : oklch(1 0 0 / 10%);
+  --input : oklch(1 0 0 / 15%);
+  --ring : oklch(0.556 0 0);
+  --chart-1 : oklch(0.488 0.243 264.376);
+  --chart-2 : oklch(0.696 0.17 162.48);
+  --chart-3 : oklch(0.769 0.188 70.08);
+  --chart-4 : oklch(0.627 0.265 303.9);
+  --chart-5 : oklch(0.645 0.246 16.439);
+  --sidebar : oklch(0.205 0 0);
+  --sidebar-foreground : oklch(0.985 0 0);
+  --sidebar-primary : oklch(0.488 0.243 264.376);
+  --sidebar-primary-foreground : oklch(0.985 0 0);
+  --sidebar-accent : oklch(0.269 0 0);
+  --sidebar-accent-foreground : oklch(0.985 0 0);
+  --sidebar-border : oklch(1 0 0 / 10%);
+  --sidebar-ring : oklch(0.556 0 0);
+}
+
+@layer base {
+  * {
+    @apply border-border outline-ring/50;
+  }
+  body {
+    @apply bg-background text-foreground;
+  }
+}
