+++ packages/ui/src/index.ts
@@ -1,0 +1,2 @@
+export { default } from "../tailwind.preset";
+export * from "./lib/utils";
