+++ packages/ui/src/shadcn/dialog.tsx
@@ -1,0 +1,143 @@
+"use client"
+
+import * as React from "react"
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+import { XIcon } from "lucide-react"
+
+import { cn } from "../lib/utils"
+
+function Dialog({
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Root>) {
+  return <DialogPrimitive.Root data-slot="dialog" {...props} />
+}
+
+function DialogTrigger({
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Trigger>) {
+  return <DialogPrimitive.Trigger data-slot="dialog-trigger" {...props} />
+}
+
+function DialogPortal({
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Portal>) {
+  return <DialogPrimitive.Portal data-slot="dialog-portal" {...props} />
+}
+
+function DialogClose({
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Close>) {
+  return <DialogPrimitive.Close data-slot="dialog-close" {...props} />
+}
+
+function DialogOverlay({
+  className,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Overlay>) {
+  return (
+    <DialogPrimitive.Overlay
+      data-slot="dialog-overlay"
+      className={cn(
+        "ui-:data-[state=open]:animate-in ui-:data-[state=closed]:animate-out ui-:data-[state=closed]:fade-out-0 ui-:data-[state=open]:fade-in-0 ui-:fixed ui-:inset-0 ui-:z-50 ui-:bg-black/50",
+        className
+      )}
+      {...props}
+    />
+  )
+}
+
+function DialogContent({
+  className,
+  children,
+  showCloseButton = true,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Content> & {
+  showCloseButton?: boolean
+}) {
+  return (
+    <DialogPortal data-slot="dialog-portal">
+      <DialogOverlay />
+      <DialogPrimitive.Content
+        data-slot="dialog-content"
+        className={cn(
+          "ui-:bg-background ui-:data-[state=open]:animate-in ui-:data-[state=closed]:animate-out ui-:data-[state=closed]:fade-out-0 ui-:data-[state=open]:fade-in-0 ui-:data-[state=closed]:zoom-out-95 ui-:data-[state=open]:zoom-in-95 ui-:fixed ui-:top-[50%] ui-:left-[50%] ui-:z-50 ui-:grid ui-:w-full ui-:max-w-[calc(100%-2rem)] ui-:translate-x-[-50%] ui-:translate-y-[-50%] ui-:gap-4 ui-:rounded-lg ui-:border ui-:p-6 ui-:shadow-lg ui-:duration-200 ui-:sm:max-w-lg",
+          className
+        )}
+        {...props}
+      >
+        {children}
+        {showCloseButton && (
+          <DialogPrimitive.Close
+            data-slot="dialog-close"
+            className="ui-:ring-offset-background ui-:focus:ring-ring ui-:data-[state=open]:bg-accent ui-:data-[state=open]:text-muted-foreground ui-:absolute ui-:top-4 ui-:right-4 ui-:rounded-xs ui-:opacity-70 ui-:transition-opacity ui-:hover:opacity-100 ui-:focus:ring-2 ui-:focus:ring-offset-2 ui-:focus:outline-hidden ui-:disabled:pointer-events-none ui-:[&_svg]:pointer-events-none ui-:[&_svg]:shrink-0 ui-:[&_svg:not([class*=size-])]:size-4"
+          >
+            <XIcon />
+            <span className="ui-:sr-only">Close</span>
+          </DialogPrimitive.Close>
+        )}
+      </DialogPrimitive.Content>
+    </DialogPortal>
+  )
+}
+
+function DialogHeader({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="dialog-header"
+      className={cn("ui-:flex ui-:flex-col ui-:gap-2 ui-:text-center ui-:sm:text-left", className)}
+      {...props}
+    />
+  )
+}
+
+function DialogFooter({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="dialog-footer"
+      className={cn(
+        "ui-:flex ui-:flex-col-reverse ui-:gap-2 ui-:sm:flex-row ui-:sm:justify-end",
+        className
+      )}
+      {...props}
+    />
+  )
+}
+
+function DialogTitle({
+  className,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Title>) {
+  return (
+    <DialogPrimitive.Title
+      data-slot="dialog-title"
+      className={cn("ui-:text-lg ui-:leading-none ui-:font-semibold", className)}
+      {...props}
+    />
+  )
+}
+
+function DialogDescription({
+  className,
+  ...props
+}: React.ComponentProps<typeof DialogPrimitive.Description>) {
+  return (
+    <DialogPrimitive.Description
+      data-slot="dialog-description"
+      className={cn("ui-:text-muted-foreground ui-:text-sm", className)}
+      {...props}
+    />
+  )
+}
+
+export {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+}
