+++ packages/ui/tailwind.config.ts
@@ -1,0 +1,9 @@
+import type { Config } from "tailwindcss";
+import preset from "./tailwind.preset";
+
+const config: Pick<Config, "prefix" | "presets"> = {
+  prefix: "ui-",
+  presets: [preset],
+};
+
+export default config;
