+++ packages/ui/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "moduleResolution": "bundler",
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["./src/*"],
+      "@hebo/ui/tailwind.preset": ["./tailwind.preset.ts"]
+    }
+  },
+  "include": ["src", "*.ts", "*.mjs"],
+  "exclude": ["node_modules", "dist"]
+}
