+++ tsconfig.base.json
@@ -14,11 +14,12 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": true,
+    "incremental": false,
     "baseUrl": ".",
     "paths": {
       "@/*": ["*"],
-      "@hebo/": ["*"],
+      "@hebo/*": ["packages/*"],
+      "@hebo/ui": ["packages/ui/"],
       "@hebo/db": ["packages/db/"],
       "@hebo/db/*": ["packages/db/*/"]
     }
